Update Vulnerability type in types
