+++ frontend/src/types/index.ts
@@ -1,14 +1,14 @@
 export * from './cpe';
 export * from './cve';
 export * from './domain';
-export * from './vulnerability';
-export * from './scan';
-export * from './organization';
-export * from './user';
-export * from './query';
-export * from './role';
-export * from './scan-task';
-export * from './saved-search';
-export * from './stats';
-export * from './report';
 export * from './notification';
+export * from './organization';
+export * from './org-query';
+export * from './query';
+export * from './report';
+export * from './role';
+export * from './saved-search';
+export * from './scan-task';
+export * from './scan';
+export * from './stats';
+export * from './user';
