--- frontend/src/types/vulnerability.ts
@@ -1,38 +1,0 @@
-import { Domain, Service } from './domain';
-
-export interface Vulnerability {
-  id: string;
-  domain: Domain;
-  createdAt: string;
-  lastSeen: string | null;
-  title: string;
-  cve: string | null;
-  isKev?: string;
-  cwe: string | null;
-  cpe: string | null;
-  description: string;
-  cvss: number | null;
-  severity: string | null;
-  state: string;
-  source: string;
-  structuredData: { [x: string]: any };
-  substate: string;
-  notes: string;
-  actions: {
-    type: 'state-change' | 'comment';
-    state?: string;
-    substate?: string;
-    value?: string;
-    automatic: boolean;
-    userId: string | null;
-    userName: string | null;
-    date: string;
-  }[];
-  references: {
-    url: string;
-    name: string;
-    source: string;
-    tags: string[];
-  }[];
-  service: Service;
-}
